// src/interfaces/references.ts
export interface ICustomerPortfolio {
  name: string;
  logo: string;
}

export interface IClientTestimonials {
  name: string;
  company: string;
  image: string;
  quote: string;
}

export interface IFeaturedProjects {
  tags: string[];
  title: string;
  description: string;
  image: string;
}

export interface ISectoralRecognition {
    title: string;
    description: string;
  }
  